Tests for Tabs component, related component tests corrections, change components structure to atomic design
