+++ app/components/organisms/DetailView/styled.js
