+++ app/components/organisms/DetailView/tests/index.test.js
@@ -11,14 +11,11 @@ describe('<DetailView />', () => {
     expect(renderedComponent.find('LatestPanelWrapper').length).toEqual(1);
   });
 
-  it('should render two tabs for table and code', () => {
+  it('should render <Tabs/>', () => {
     const renderedComponent = shallow(
       <DetailView data={{}} fields={{}} loading={false} error={false} />,
     );
-
-    expect(renderedComponent.find('LatestPanelWrapper').length).toEqual(1);
-    expect(renderedComponent.find('ListItem').length).toEqual(2);
-    expect(renderedComponent.find('ListItem Button').length).toEqual(2);
+    expect(renderedComponent.find('Tabs').length).toEqual(1);
   });
 
   it('should render <LoadingIndicator /> when fetching data is in progress', () => {
@@ -62,6 +59,7 @@ describe('<DetailView />', () => {
 
     renderedComponent.setState({ selectedTabId: 'code' });
     expect(renderedComponent.find('SyntaxHighlighter').length).toEqual(1);
+    expect(renderedComponent.find('Table').length).toEqual(0);
   });
 
   it('should render table view when table tab is selected', () => {
@@ -71,59 +69,6 @@ describe('<DetailView />', () => {
 
     renderedComponent.setState({ selectedTabId: 'table' });
     expect(renderedComponent.find('SyntaxHighlighter').length).toEqual(0);
-    expect(renderedComponent.find('ul').length).toEqual(1);
-  });
-
-  it('should assign `active` class to clicked element', () => {
-    const renderedComponent = shallow(
-      <DetailView data={{}} fields={{}} loading={false} error={false} />,
-    );
-
-    expect(
-      renderedComponent
-        .find('ListItem Button')
-        .at(0)
-        .hasClass('active'),
-    ).toEqual(true);
-    expect(
-      renderedComponent
-        .find('ListItem Button')
-        .at(1)
-        .hasClass('active'),
-    ).toEqual(false);
-
-    renderedComponent
-      .find('ListItem Button')
-      .at(1)
-      .simulate('click');
-    expect(
-      renderedComponent
-        .find('ListItem Button')
-        .at(0)
-        .hasClass('active'),
-    ).toEqual(false);
-    expect(
-      renderedComponent
-        .find('ListItem Button')
-        .at(1)
-        .hasClass('active'),
-    ).toEqual(true);
-
-    renderedComponent
-      .find('ListItem Button')
-      .at(0)
-      .simulate('click');
-    expect(
-      renderedComponent
-        .find('ListItem Button')
-        .at(0)
-        .hasClass('active'),
-    ).toEqual(true);
-    expect(
-      renderedComponent
-        .find('ListItem Button')
-        .at(1)
-        .hasClass('active'),
-    ).toEqual(false);
+    expect(renderedComponent.find('Table').length).toEqual(1);
   });
 });
