+++ app/containers/AccountPage/index.js
@@ -15,8 +15,8 @@ import { FormattedMessage, intlShape } from 'react-intl';
 import config from 'config';
 import injectSaga from 'utils/injectSaga';
 import injectReducer from 'utils/injectReducer';
-import DetailView from 'components/DetailView';
-import ListView from 'components/ListView';
+import DetailView from 'components/organisms/DetailView';
+import ListView from 'components/organisms/ListView';
 import TransactionAddressLink from 'components/TransactionAddressLink';
 import { makeSelectAccount, makeSelectTransactions } from './selectors';
 import reducer from './reducer';
