+++ app/containers/BlockPage/styled.js
@@ -1,6 +1,6 @@
 /* eslint-disable prettier/prettier */
 import styled from 'styled-components';
-import { LatestPanelWrapper } from '../../components/LatestPanel/styled';
+import { LatestPanelWrapper } from '../../components/organisms/LatestPanel/styled';
 export const BlockPageWrapper = styled.div`
   /* https://github.com/styled-components/stylelint-processor-styled-components/issues/54 */
   ${/* sc-selector */LatestPanelWrapper} {
