+++ app/containers/MessagePage/index.js
@@ -15,8 +15,8 @@ import { Link } from 'react-router-dom';
 import config from 'config';
 import injectSaga from 'utils/injectSaga';
 import injectReducer from 'utils/injectReducer';
-import DetailView from 'components/DetailView';
-import ListView from 'components/ListView';
+import DetailView from 'components/organisms/DetailView';
+import ListView from 'components/organisms/ListView';
 import TransactionAddressLink from 'components/TransactionAddressLink';
 import { makeSelectMessage, makeSelectTransactions } from './selectors';
 import reducer from './reducer';
