+++ app/containers/TransactionsListPage/index.js
@@ -15,7 +15,7 @@ import { Link } from 'react-router-dom';
 import injectSaga from 'utils/injectSaga';
 import injectReducer from 'utils/injectReducer';
 import config from 'config';
-import ListView from 'components/ListView';
+import ListView from 'components/organisms/ListView';
 import TransactionAddressLink from 'components/TransactionAddressLink';
 import { makeSelectTransactions } from './selectors';
 import reducer from './reducer';
